# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CRM Test Documentation",
        "description": "Docmentation for CRM Test",
        "contact": {
            "email": "gilaarief@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "CRM Test Server"
        }
    ],
    "paths": {
...
        "/api/companies": {
            "get": {
                "tags": [
                    "Company"
                ],
                "summary": "Get all companies",
                "operationId": "a261a49d113a2af2a5ccb5798c6aae7e",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search keyword to filter companies.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "description": "Column to sort companies by (default: name).",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "orderDirection",
                        "in": "query",
                        "description": "Sort direction (asc or desc, default: asc).",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of companies",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/company": {
            "post": {
                "tags": [
                    "Company"
                ],
                "summary": "Create a new company",
                "operationId": "ac33eaa8422bb09ea6dfe7fe24255ed0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phone"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Company's name",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Unique email for the user",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": "Phone number (10-16 digits)",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Company created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "Validation errors occurred.",
                                    "errors": {
                                        "name": "The name field is required.",
                                        "email": "The email field is required and must be a valid email address.",
                                        "phone": "The phone number must be between 10 and 16 digits."
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/company/{id}": {
            "get": {
                "tags": [
                    "Company"
                ],
                "summary": "Get Company details",
                "operationId": "0c96d2d82fb75301c38a534ab05e61d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Company",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Company details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/company/{id}": {
            "put": {
                "tags": [
                    "Company"
                ],
                "summary": "Update Company details",
                "operationId": "616ef40178c2fbc1396e55aeefea3809",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Company to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phone"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Company's name",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Company's email",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": "Company's phone number",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "Validation errors occurred.",
                                    "errors": {
                                        "name": "The name field is required.",
                                        "email": "The email field is required and must be a valid email address.",
                                        "phone": "The phone number must be between 10 and 16 digits."
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Company"
                ],
                "summary": "Delete a Company",
                "operationId": "f54ecf7112ff4ce4b555e90e5695d10f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Company to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Company deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "jwt",
                "scheme": "bearer"
            }
        }
    },
...
}